Moved select on clickj and stop event in separate directives
