+++ client/app/scripts/liveblog-edit/directives/select-text-on-click.js
@@ -1,0 +1,10 @@
+export default function selectTextOnClick() {
+    return {
+        link: function(scope, elem, attrs) {
+            elem.bind('click', function() {
+                elem.focus();
+                elem.select();
+            });
+        }
+    };
+}
