+++ client/app/scripts/liveblog-edit/directives/stop-event.js
@@ -1,0 +1,10 @@
+export default function stopEvent() {
+    return {
+        restrict: 'A',
+        link: function (scope, element, attr) {
+            element.bind(attr.stopEvent, function (e) {
+                e.stopPropagation();
+            });
+        }
+    };
+}
